requirements.txt
aiohttp
beautifulsoup4
httpx
mcp
openai
pandas
pypdf
python-docx
python-dotenv
python-jobspy
slack_bolt
streamlit
